+++ dftt_timecode/__init__.py
@@ -2,4 +2,5 @@ name = 'dftt_timecode'
 __author__ = 'You Ziyuan'
 __version__ = '0.0.9'
 
-from dftt_timecode.core import DfttTimecode
+from dftt_timecode.core.dftt_timecode import DfttTimecode
+from dftt_timecode.core.dftt_timecode import DfttTimeRange